BAU Setup logger and log don't throw when no VCs found
Setup logger and add info logs with context. Also in the case when no VCs are found
for the given user id, log rather than throwing an error - because this is expected
behaviour for users who haven't been through IPV.

+++ lambdas/delete-user-data/src/delete-data.ts
@@ -1,19 +1,23 @@
 import { DynamoDBClient } from "@aws-sdk/client-dynamodb";
 import { DynamoDBDocument } from "@aws-sdk/lib-dynamodb";
 import { config } from "./config";
+import { logger } from "./logger";
 import { VCItemKey } from "./types";
 
 const ddbDocClient = DynamoDBDocument.from(new DynamoDBClient({ region: "eu-west-2" }));
 
 export const deleteVCs = async (userId: string): Promise<void> => {
   const keysToDelete = await getVCItemKeys(userId);
-  const deletePromises = keysToDelete.map((key) =>
-    ddbDocClient.delete({
-      TableName: config.userIssuedCredentialsTableName,
-      Key: key,
-    })
-  );
-  await Promise.all(deletePromises);
+  if (keysToDelete.length > 0) {
+    const deletePromises = keysToDelete.map((key) =>
+      ddbDocClient.delete({
+        TableName: config.userIssuedCredentialsTableName,
+        Key: key,
+      })
+    );
+    await Promise.all(deletePromises);
+    logger.info("Deleted user's VCs", { userId, count: keysToDelete.length });
+  }
 };
 
 const getVCItemKeys = async (userId: string): Promise<VCItemKey[]> => {
@@ -26,7 +30,9 @@ const getVCItemKeys = async (userId: string): Promise<VCItemKey[]> => {
     },
   });
   if (!itemKeys || itemKeys.length === 0) {
-    throw new Error("failed to find any VCs for user: " + userId);
+    logger.info("No VCs found for user", { userId });
+    return [];
   }
+  logger.info("Found user's VCs", { userId, count: itemKeys.length });
   return itemKeys as VCItemKey[];
 };
